\name{aws.cloudtrail-package}
\alias{aws.cloudtrail-package}
\alias{aws.cloudtrail}
\docType{package}
\title{AWS CloudTrail Client Package}
\description{A simple client package for the Amazon Web Services AWS CloudTrail REST API.}
\details{}
\author{Thomas J. Leeper <thosjleeper@gmail.com>}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}
