% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trails.R
\name{get_trails}
\alias{get_trails}
\title{get_trails}
\usage{
get_trails(...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{cloudtrailHTTP}}.}
}
\value{
A list of objects of class \dQuote{aws_cloudtrail}.
}
\description{
Get list of trails
}
\details{
Get a list of available cloudtrails.
}
\examples{
\dontrun{
get_trails()
}
}
\references{
\url{http://docs.aws.amazon.com/awscloudtrail/latest/APIReference/API_DescribeTrails.html}
}
\seealso{
\code{\link{create_trail}}
}
