% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudtrail.R
\name{start_logging}
\alias{start_logging}
\alias{stop_logging}
\title{logging}
\usage{
start_logging(name, ...)

stop_logging(name, ...)
}
\arguments{
\item{name}{A character string specifying the name of a Cloudtrail or the ARN for a Cloudtrail.}

\item{\dots}{Additional arguments passed to \code{\link{cloudtrailHTTP}}.}
}
\description{
Start/stop logging
}
\details{
\code{start_logging} starts a
}
\examples{
\dontrun{
trail <- create_trail("exampletrail", "mys3bucket")
start_logging("exampletrail")
stop_logging("exampletrail")
trail_status("exampletrail")
delete_trail("exampletrail")
}
}
\references{
\url{http://docs.aws.amazon.com/awscloudtrail/latest/APIReference/API_StartLogging.html}, \url{http://docs.aws.amazon.com/awscloudtrail/latest/APIReference/API_StopLogging.html}
}
\seealso{
\code{\link{create_trail}}, \code{\link{trail_status}}
}

