% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.cloudtrail-package.R
\docType{package}
\name{aws.cloudtrail-packaage}
\alias{aws.cloudtrail-packaage}
\alias{aws.cloudtrail-package}
\alias{aws.cloudtrail}
\alias{aws.cloudtrail-packaage-package}
\title{aws.cloudtrail}
\description{
AWS CloudTrail Client Package
}
\details{
A simple client package for the Amazon Web Services AWS CloudTrail REST API.
}
\references{
\url{http://docs.aws.amazon.com/awscloudtrail/latest/userguide/cloudtrail-user-guide.html}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
