% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{cloudtrailHTTP}
\alias{cloudtrailHTTP}
\title{cloudtrailHTTP}
\usage{
cloudtrailHTTP(query, headers = list(), body = NULL,
  version = "2013-11-01", verbose = getOption("verbose", FALSE),
  region = Sys.getenv("AWS_DEFAULT_REGION", "us-east-1"), key = NULL,
  secret = NULL, session_token = NULL, ...)
}
\arguments{
\item{query}{A list.}

\item{headers}{A list of headers to pass to the HTTP request.}

\item{body}{The body of the request.}

\item{version}{A character string specifying the API version.}

\item{verbose}{A logical indicating whether to be verbose. Default is given by \code{options("verbose")}.}

\item{region}{A character string containing the AWS region. If missing, defaults to \dQuote{us-east-1}.}

\item{key}{A character string containing an AWS Access Key ID. See \code{\link[aws.signature]{locate_credentials}}.}

\item{secret}{A character string containing an AWS Secret Access Key. See \code{\link[aws.signature]{locate_credentials}}.}

\item{session_token}{A character string containing an AWS Session Token. See \code{\link[aws.signature]{locate_credentials}}.}

\item{\dots}{Additional arguments passed to \code{\link[httr]{POST}}.}
}
\value{
A list.
}
\description{
AWS Cloudtrail API Requests
}
\details{
This is the workhorse function for executing Cloudtrail API requests. It is typically not necessary to call this function directly.
}
