% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{cloudtrailHTTP}
\alias{cloudtrailHTTP}
\title{cloudtrailHTTP}
\usage{
cloudtrailHTTP(query, body = NULL, headers = list(),
  version = "2013-11-01", region = Sys.getenv("AWS_DEFAULT_REGION",
  "us-east-1"), key = NULL, secret = NULL, session_token = NULL, ...)
}
\arguments{
\item{query}{A list.}

\item{body}{The body of the request.}

\item{headers}{A list of headers to pass to the REST request.}

\item{version}{A character string specifying the API version.}

\item{region}{A character string containing the AWS region. If missing, defaults to the value of environment variable \samp{AWS_DEFAULT_REGION}.}

\item{key}{A character string containing an AWS Access Key ID. If missing, defaults to value stored in environment variable \dQuote{AWS_ACCESS_KEY_ID}.}

\item{secret}{A character string containing an AWS Secret Access Key. If missing, defaults to value stored in environment variable \dQuote{AWS_SECRET_ACCESS_KEY}.}

\item{session_token}{Optionally, a character string containing an AWS temporary Session Token. If missing, defaults to value stored in environment variable \dQuote{AWS_SESSION_TOKEN}.}

\item{\dots}{Additional arguments passed to \code{\link[httr]{POST}}.}
}
\value{
A list.
}
\description{
AWS Cloudtrail API Requests
}
\details{
This is the workhorse function for executing Cloudtrail API requests. It is typically not necessary to call this function directly.
}
