% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trail_status.R
\name{trail_status}
\alias{trail_status}
\title{trail_status}
\usage{
trail_status(name, ...)
}
\arguments{
\item{name}{A character string specifying the name of a Cloudtrail or the ARN for a Cloudtrail, or an object of class \dQuote{aws_cloudtrail}.}

\item{\dots}{Additional arguments passed to \code{\link{cloudtrailHTTP}}.}
}
\value{
A list.
}
\description{
Get trail status/history
}
\details{
This function returns full details and history for a trail, including errors and logging start/stop times.
}
\references{
\url{http://docs.aws.amazon.com/awscloudtrail/latest/APIReference/API_GetTrailStatus.html}
}
\seealso{
\code{\link{get_trails}}, \code{\link{start_logging}}, \code{\link{create_trail}}
}
